+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
 ## Unreleased changes
+
+- Fix a bug where the block state hash was not returned properly for the genesis block.
 - Do not reset banned peers on startup by default.
 - The node remembers peers across restarts. When starting up it will try to connect to stored peers in addition to any supplied bootstrap and given nodes. 
   Use the new flag `--clear-persisted-peers` (environment variable  `CONCORDIUM_NODE_CLEAR_PERSISTED_PEERS`) to clear stored peers on startup.
 - Renamed the flag `--no-clear-bans` to `clear-bans`. When set it will clear the banned peers on startup.
+
 
 - Fix a bug where credential registration IDs for genesis accounts were not
   correctly recorded. As a result, the index of accounts by credential ids was